% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_methods.R
\name{ankan}
\alias{ankan}
\title{Compute the p-value for conditional independence using the method of Ankan and Textor 2022.}
\usage{
ankan(x, y, f, g)
}
\arguments{
\item{x}{integer vector of length n taking values in {1, ..., dx}.}

\item{y}{integer vector of length n taking values in {1, ..., dy}.}

\item{f}{matrix of dimension n by dx. ith row is the estimate of (Pr(X=1|Zi), ..., Pr(X=dx|Zi)).}

\item{g}{matrix of dimension n by dy. ith row is the estimate of (Pr(Y=1|Zi), ..., Pr(Y=dy|Zi)).}
}
\value{
float p-value.
}
\description{
Compute the p-value for conditional independence using the method of Ankan and Textor 2022.
}
