% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merging_functions.R
\name{approx_chi_metric}
\alias{approx_chi_metric}
\title{Evaluate the approximate chi-squared cdf using the method of Box (1954).}
\usage{
approx_chi_metric(normsq, tr, tr2)
}
\arguments{
\item{normsq}{float equal to \|T_vector\|^2_2.}

\item{tr}{float equal to tr(Sigma).}

\item{tr2}{float equal to tr(Sigma^2).}
}
\value{
float between 0 and 1
}
\description{
Under the null T is asymptotically N(0, Sigma), and \|T\|_2^2 can be
approximated by a rescaled chi-square with scaling and degrees of freedom
determined by Sigma.
}
