% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_functions.R
\name{bootstrap_T}
\alias{bootstrap_T}
\title{Generate bootstrap samples T ~ N(0, Sigma)}
\usage{
bootstrap_T(Sigma, n_boot)
}
\arguments{
\item{Sigma}{covariance matrix with dimension dx\emph{dy by dx}dy.}

\item{n_boot}{integer number of bootstrap samples.}
}
\value{
dx*dy by n_boot matrix, each column of which is distributed as N(0,Sigma).
}
\description{
Generate bootstrap samples T ~ N(0, Sigma)
}
