% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_settings.R
\name{evaluate_sim}
\alias{evaluate_sim}
\title{Wrapper function to evaluate simulation data.}
\usage{
evaluate_sim(data, dx, dy, n_boot, methods)
}
\arguments{
\item{data}{list containing T_vector and Sigma.}

\item{dx}{integer number of x labels.}

\item{dy}{integer number of y labels.}

\item{n_boot}{integer number of bootstrap samples.}

\item{methods}{vector of strings specifying the query functions.}
}
\value{
vector of p-values corresponding to different methods.
}
\description{
Wrapper function to evaluate simulation data.
}
