% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_settings.R
\name{form_T_Sigma}
\alias{form_T_Sigma}
\title{Form T_vector and Sigma from observed labels x and y and probability estimate
matrices f and g.}
\usage{
form_T_Sigma(x, y, f, g, normalise)
}
\arguments{
\item{x}{integer vector of length n taking values in {1, ..., dx}.}

\item{y}{integer vector of length n taking values in {1, ..., dy}.}

\item{f}{matrix of dimension n by dx. ith row is the estimate of (Pr(X=1|Zi), ..., Pr(X=dx|Zi)).}

\item{g}{matrix of dimension n by dy. ith row is the estimate of (Pr(Y=1|Zi), ..., Pr(Y=dy|Zi)).}

\item{normalise}{logical determining if normalised or unnormalised generalised covariance is to be used.}
}
\value{
list containing T_vector and Sigma matrix.
}
\description{
Form T_vector and Sigma from observed labels x and y and probability estimate
matrices f and g.
}
