% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{greedy_query}
\alias{greedy_query}
\title{Compute metrics after successive label merging.}
\usage{
greedy_query(
  T_vector,
  Sigma,
  dx,
  dy,
  metric,
  xsearch,
  ysearch,
  colsample_bylevel,
  trees = list(NULL, NULL)
)
}
\arguments{
\item{T_vector}{vector of length dx*dy.}

\item{Sigma}{matrix of dimension dx\emph{dy by dx}dy.}

\item{dx}{integer dimension of X.}

\item{dy}{integer dimension of Y.}

\item{metric}{str "approx_chi" or "norm" controlling the metric used for searching.}

\item{xsearch}{str "greedy" or "ordinal" controlling the X search procedure.}

\item{ysearch}{str "greedy" or "ordinal" controlling the Y search procedure.}

\item{colsample_bylevel}{float in (0,1]. The subsample ratio of labels considered for merging at each level.
Subsampling occurs once for every new depth level reached.}

\item{trees}{list containing tree structures (only used if xsearch or ysearch == tree)}
}
\value{
list containing vector of metrics and the corresponding merged labels.
}
\description{
Merge pairs of labels for each of X and Y so as to maximise the
metric at each stage.
}
