% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_methods.R
\name{Li_residual}
\alias{Li_residual}
\title{Compute the residual of Li and Shepherd 2012.}
\usage{
Li_residual(x, f)
}
\arguments{
\item{x}{integer vector of length n taking values in {1, ..., dx}.}

\item{f}{matrix of dimension n by dx. ith row is the estimate of (Pr(X=1|Zi), ..., Pr(X=dx|Zi)).}
}
\value{
vector of residuals.
}
\description{
Compute the residual of Li and Shepherd 2012.
}
