% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_methods.R
\name{multinomial}
\alias{multinomial}
\title{Compute the p-value for conditional independence using nested multinomial models.}
\usage{
multinomial(x, y, z)
}
\arguments{
\item{x}{integer vector of length n taking values in {1, ..., dx}.}

\item{y}{integer vector of length n taking values in {1, ..., dy}.}

\item{z}{matrix with n rows.}
}
\value{
float p-value.
}
\description{
Compute the p-value for conditional independence using nested multinomial models.
}
