% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merging_functions.R
\name{update_normsq}
\alias{update_normsq}
\title{Update \|T\|^2_2 after merging pairs of indices.}
\usage{
update_normsq(normsq, T_vector, index1, index2)
}
\arguments{
\item{normsq}{float equal to \|T_vector\|^2_2.}

\item{T_vector}{vector of length p.}

\item{index1}{boolean vector of length p, with d TRUE entries.}

\item{index2}{boolean vector of length p, with d TRUE entries.}
}
\value{
float equal to the squared norm of T after merging the pairs of
labels.
}
\description{
Update \|T\|^2_2 after merging pairs of indices.
}
