% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merging_functions.R
\name{update_Sigma}
\alias{update_Sigma}
\title{Update Sigma matrix by merging pairs of indices.}
\usage{
update_Sigma(Sigma, index1, index2)
}
\arguments{
\item{Sigma}{matrix of dimension p by p.}

\item{index1}{boolean vector of length p, with d TRUE entries.}

\item{index2}{boolean vector of length p, with d TRUE entries.}
}
\value{
matrix of dimension (p-d) by (p-d) corresponding to merging the d
pairs of labels indexed by index1 and index2.
}
\description{
Update Sigma matrix by merging pairs of indices.
}
