% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merging_functions.R
\name{update_T}
\alias{update_T}
\title{Update T vector by merging pairs of indices.}
\usage{
update_T(T_vector, index1, index2)
}
\arguments{
\item{T_vector}{vector of length p.}

\item{index1}{boolean vector of length p, with d TRUE entries.}

\item{index2}{boolean vector of length p, with d TRUE entries.}
}
\value{
vector of length p-d corresponding to merging the d pairs of
labels indexed by index1 and index2.
}
\description{
Update T vector by merging pairs of indices.
}
