% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merging_functions.R
\name{update_tr}
\alias{update_tr}
\title{Update tr(Sigma) after merging pairs of indices.}
\usage{
update_tr(tr, Sigma, index1, index2)
}
\arguments{
\item{tr}{float equal to tr(Sigma).}

\item{Sigma}{matrix of dimension p by p.}

\item{index1}{boolean vector of length p, with d TRUE entries.}

\item{index2}{boolean vector of length p, with d TRUE entries.}
}
\value{
float equal to the trace of Sigma after merging the pairs of
labels.
}
\description{
Update tr(Sigma) after merging pairs of indices.
}
