% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merging_functions.R
\name{update_tr2}
\alias{update_tr2}
\title{Update tr(Sigma^2) after merging pairs of indices.}
\usage{
update_tr2(tr2, Sigma, index1, index2)
}
\arguments{
\item{tr2}{float equal to tr(Sigma^2).}

\item{Sigma}{matrix of dimension p by p.}

\item{index1}{boolean vector of length p, with d TRUE entries.}

\item{index2}{boolean vector of length p, with d TRUE entries.}
}
\value{
float equal to the trace of Sigma^2 after merging the pairs of
labels.
}
\description{
Update tr(Sigma^2) after merging pairs of indices.
}
