% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merging_functions.R
\name{X_index}
\alias{X_index}
\title{Return the indices whose x-entry corresponds to element j.}
\usage{
X_index(j, dx, dy)
}
\arguments{
\item{j}{integer index 1 <= j <= dx.}

\item{dx}{integer dimension of X.}

\item{dy}{integer dimension of Y.}
}
\value{
boolean vector of length dx*dy with kth TRUE entry corresponding
to index (j,k).
}
\description{
For a vector of length dx*dy indexed (1,1), ..., (dx, 1), (1, 2), ...,
(dx, 2), ..., (1, dy), ..., (dx, dy), this function returns the indices
corresponding to (j,k) for k = 1, ..., dy.
}
