% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merging_functions.R
\name{Y_index}
\alias{Y_index}
\title{Return the indices whose y-entry corresponds to element k.}
\usage{
Y_index(k, dx, dy)
}
\arguments{
\item{k}{integer index 1 <= k <= dy.}

\item{dx}{integer dimension of X.}

\item{dy}{integer dimension of Y.}
}
\value{
boolean vector of length dx*dy with jth TRUE entry corresponding
to index (j,k).
}
\description{
For a vector of length dx*dy indexed (1,1), ..., (dx, 1), (1, 2), ...,
(dx, 2), ..., (1, dy), ..., (dx, dy), this function returns the indices
corresponding to (j,k) for j = 1, ..., dx.
}
